+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - added tenant prediction report in myems-web
 - added shopfloor prediction report in myems-web
 - added store prediction report in myems-web
+- added the function of displaying the password entered by the user to myems-admin
 ### Changed
 - changed node.js version to 22.22.2, 24.5.0 or later in myems-web
 - changed npm version to 10.9.7, 11.5.1 or later in myems-web